requirements.txt
anyio==3.7.0
APScheduler==3.10.1
Authlib==0.11
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.160
botocore==1.29.160
bs4==0.0.1
cachetools==4.2.4
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
distlib==0.3.6
dnspython==2.3.0
docopt==0.6.2
et-xmlfile==1.1.0
fastapi==0.98.0
filelock==3.12.2
Flask==1.1.2
Flask-Login==0.4.1
google-api-core==2.11.1
google-api-python-client==2.90.0
google-auth==1.27.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.59.1
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.22.0
idna==2.8
ischedule==1.2.6
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.3
jmespath==1.0.1
MarkupSafe==2.1.3
oauthlib==3.1.0
openpyxl==2.4.11
packaging==23.1
pathlib==1.0.1
pipdeptree==2.9.3
platformdirs==3.8.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pymongo==3.11.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
records==0.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.9
s3transfer==0.6.1
schedule==1.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
starlette==0.27.0
tablib==3.5.0
typing_extensions==4.6.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
vendor==0.1
virtualenv==20.23.1
waitress==2.1.2
Werkzeug==2.0.3
